order is devilered functionality completed
